+++ app/src/main/java/ru/ds/materialdesign/view/ux/ButtonUXFragment.kt
@@ -7,28 +7,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.ds.materialdesign.databinding.FragmentUxButtonBinding
 
-class ButtonUXFragment : Fragment() {
-    private var _binding: FragmentUxButtonBinding? = null
-    private val binding: FragmentUxButtonBinding
-        get() = _binding!!
+class ButtonUXFragment : ViewBindingFragment<FragmentUxButtonBinding>(FragmentUxButtonBinding::inflate) {
 
     companion object{
         fun newInstance() = ButtonUXFragment()
     }
 
-
-    override fun onCreateView(
-            inflater: LayoutInflater,
-            container: ViewGroup?,
-            savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentUxButtonBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
 }
